Renaming ANSI_COMMAND.hpp to COMMAND.hpp and moving ANSI files to an ANSI folder.
