+++ src/tests/basic_test.cpp
@@ -1,4 +1,4 @@
-#include "../ANSI.hpp"
+#include "../ANSI/ANSI.hpp"
 
 #include <iostream>
